src/src/test/java/com/trade_accounting/controllers/rest/finance/LossRestControllerTest.java
package src.test.java.com.trade_accounting.controllers.rest.finance;


import com.google.gson.Gson;
import com.trade_accounting.controllers.rest.finance.LossRestController;
import com.trade_accounting.models.dto.finance.LossDto;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.MediaType;
import org.springframework.restdocs.mockmvc.MockMvcRestDocumentation;
import org.springframework.security.test.context.support.WithUserDetails;
import org.springframework.test.context.TestPropertySource;
import org.springframework.test.context.jdbc.Sql;
import org.springframework.test.web.servlet.MockMvc;

import java.util.List;

import static org.hamcrest.Matchers.hasSize;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.springframework.security.test.web.servlet.response.SecurityMockMvcResultMatchers.authenticated;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@SpringBootTest
@AutoConfigureMockMvc
@TestPropertySource(properties = {"spring.config.location = src/test/resources/application-test.yml"})
@Sql(value = "/Loss-before.sql", executionPhase = Sql.ExecutionPhase.BEFORE_TEST_METHOD)
@WithUserDetails(value = "karimogon@mail.ru")
@AutoConfigureRestDocs(outputDir = "target/snippets", uriScheme = "http", uriPort = 4444)
public class LossRestControllerTest {

    @Autowired
    private LossRestController LossController;

    @Autowired
    private MockMvc mockMvc;

    @Test
    void testExistence() {
        assertNotNull(LossController, "Loss Rest Controller is null");
    }

    @Test
    void getAll() throws Exception {
        mockMvc.perform(get("/api/loss"))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(authenticated())
                .andExpect(jsonPath("$", hasSize(3)))
                .andDo(MockMvcRestDocumentation.document("{class-name}/{method-name}"));
    }

    @Test
    void getById() throws Exception {
        String lossDtoJson = new Gson().toJson(LossDto.builder()
                .id(1L)
                .date("2021-07-01 11:11")
                .warehouseId(1L)
                .companyId(1L)
                .isSent(false)
                .isPrint(true)
                .comment("Comment1")
                .build());

        mockMvc.perform(get("/api/loss/1"))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(authenticated())
                .andExpect(content().json(lossDtoJson))
                .andDo(MockMvcRestDocumentation.document("{class-name}/{method-name}"));
    }

    @Test
    void create() throws Exception {
        String lossDtoJson = new Gson().toJson(LossDto.builder()
                .date("2021-08-08 22:22")
                .warehouseId(3L)
                .companyId(3L)
                .isSent(true)
                .isPrint(true)
                .comment("Comment4")
                .lossProductsIds(List.of(1L, 2L))
                .build());

        mockMvc.perform(post("/api/loss")
                        .contentType(MediaType.APPLICATION_JSON).content(lossDtoJson))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(authenticated())
                .andExpect(content().json(lossDtoJson))
                .andDo(MockMvcRestDocumentation.document("{class-name}/{method-name}"));

        mockMvc.perform(get("/api/loss"))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(authenticated())
                .andExpect(jsonPath("$", hasSize(4)));
    }

    @Test
    void update() throws Exception {
        String lossDtoJson = new Gson().toJson(LossDto.builder()
                .date("2021-03-03 11:45")
                .warehouseId(3L)
                .companyId(3L)
                .isSent(true)
                .isPrint(true)
                .comment("CommentUpdate")
                .lossProductsIds(List.of(1L, 2L))
                .build());

        mockMvc.perform(put("/api/loss")
                        .contentType(MediaType.APPLICATION_JSON).content(lossDtoJson))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(authenticated())
                .andExpect(content().json(lossDtoJson))
                .andDo(MockMvcRestDocumentation.document("{class-name}/{method-name}"));
        mockMvc.perform(get("/api/loss"))
                .andDo(print());
    }

    @Test
    void deleteById() throws Exception {
        mockMvc.perform(delete("/api/loss/2"))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(authenticated())
                .andDo(MockMvcRestDocumentation.document("{class-name}/{method-name}"));
        mockMvc.perform(get("/api/loss"))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(authenticated())
                .andExpect(jsonPath("$", hasSize(2)));
    }


}
